started implementing users
