+++ views/inc/nav.ejs
@@ -1,0 +1,9 @@
+<nav class="navbar navbar-dark bg-dark" role="navigation">
+    <div>
+        <a class="navbar-brand" href="/">Space Objects</a>
+    </div>
+    <div>
+        <a class="navbar-brand" href="/login">Login</a>
+        <a class="navbar-brand" href="/register">Register</a>
+    </div>
+</nav>
