+++ views/layouts/app.ejs
@@ -7,11 +7,14 @@
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.3.1/css/bootstrap.min.css">
     <link rel="stylesheet" href="/resources/css/style.css">
     <script src="https://unpkg.com/jquery@3.3.1/dist/jquery.min.js"></script>
-    <script src="https://unpkg.com/bootstrap@4.3.1/dist/js/bootstrap.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
+    <script src="/resources/js/main.js"></script>
     <title>Space Objects</title>
 </head>
 <body>
-    <div class="container mt-4">
+        <%- include('../inc/nav')%>
+    <div class="container mt-4 main-container">
         <%- body %>
     </div>
 </body>
