+++ views/pages/login.ejs
@@ -1,0 +1,17 @@
+<div class="col-md-6 login-form-1">
+    <h3>Login</h3>
+    <form action="/auth" method="POST">
+        <div class="form-group">
+            <input type="text" class="form-control" placeholder="Your Email" value="" />
+        </div>
+        <div class="form-group">
+            <input type="password" class="form-control" placeholder="Your Password" value="" />
+        </div>
+        <div class="form-group">
+            <input type="submit" class="btn-info btn" value="Login" />
+        </div>
+        <div class="form-group">
+            Don't have an account? <a href="/register">Register one</a>
+        </div>
+    </form>
+</div>
