+++ views/pages/object.ejs
@@ -15,6 +15,7 @@
             --- Argument of perigee (degrees): <%= object.perigee %>
             --- Mean anomaly (degrees): <%= object.anomaly %>
             --- Eccentricity (decimal point assumed): <%= object.eccentricity %>
+            --- B Star Drag: <%= object.drag %>
         
         Revolutions per day (mean motion): <%= object.motion %>
         Revolution count at epoch: <%= object.revolution %>
