+++ views/pages/register.ejs
@@ -1,0 +1,24 @@
+<div class="col-md-6 login-form-1">
+    <h3>Login</h3>
+    <form action="/auth" method="POST">
+        <div class="form-group">
+            <input type="text" name="uname" class="form-control" placeholder="User name" id="userName">
+        </div>
+        <div class="form-group">
+            <input type="text" name="email" class="form-control" placeholder="Your Email" id="email" />
+        </div>
+        <div class="form-group">
+            <input type="password" name="password" class="form-control" placeholder="Your Password" id="password-1" />
+        </div>
+        <div class="form-group">
+            <input type="password" name="password" class="form-control" placeholder="Confirm password" id="password-2" />
+        </div>
+        <div class="form-group">
+            <input type="submit" class="btn-info btn" value="Login" id="submit"/>
+        </div>
+        <div class="form-group">
+            Already have an account? <a href="/login">Login</a>
+        </div>
+    </form>
+</div>
+<script src="/resources/js/auth.js"></script>
